server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-engineio==4.9.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.6
openai==1.10.0
